+++ CHANGELOG.md
@@ -13,11 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--
+- Improve logging in the plugin loader.
 
 ### Fixed
 
 - Movie clean title table would not properly save/update user settings or defaults after being modified once.
+- Subprocess windows executing in a new window on Windows.
 
 ### Removed